Move fakeRuleStore to tests/fakes package
